+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 Changelog
 =========
 
-## vNext-ITN1
+## vNext
+
+### Fixes
+
+- Fixed decentralization countdown storybook story ([PR 1863](https://github.com/input-output-hk/daedalus/pull/1863))
 
 ### Chores
 
@@ -24,6 +28,7 @@ Changelog
 - Implemented acceptance tests for wallet delegation ([PR 1814](https://github.com/input-output-hk/daedalus/pull/1814))
 - Improved loading state of the "Stake pools" screen ([PR 1814](https://github.com/input-output-hk/daedalus/pull/1814))
 - Updated `cardano-wallet` to version `2020-02-17` ([PR 1856](https://github.com/input-output-hk/daedalus/pull/1856))
+- Updated `cardano-wallet` to revision `573a7038` ([PR 1862](https://github.com/input-output-hk/daedalus/pull/1862))
 
 ## 2.1.0-ITN1